test(settings): target connector settings owner
